refactor(View): InputView와 OutputView로 분리
- 하나의 클래스였던 View를 입력과 출력 역할에 따라 분리
