+++ src/main/java/lotto/Application.java
@@ -1,7 +1,10 @@
 package lotto;
 
+import lotto.View.OutputView;
+
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
+        OutputView view = new OutputView();
     }
 }
