+++ src/main/java/lotto/View/OutputView.java
@@ -1,13 +1,10 @@
-package lotto;
+package lotto.View;
 
-import camp.nextstep.edu.missionutils.Console;
 import java.util.List;
+import lotto.Lotto;
 
-public class View {
+public class OutputView {
 
-    private static final String LOTTO_PURCHASE_AMOUNT_OUTPUT_STATEMENT = "구입금액을 입력해 주세요.";
-    private static final String LOTTO_WINNING_NUMBER_OUTPUT_STATEMENT = "당첨 번호를 입력해 주세요.";
-    private static final String LOTTO_BONUS_NUMBER_OUTPUT_STATEMENT = "보너스 번호를 입력해 주세요.";
     private static final String PURCHASED_LOTTO_NUMBER_OUTPUT_STATEMENT = "개를 구매했습니다.";
     private static final List<String> WINNING_DETAIL_OUTPUT_STATEMENTS = List.of(
             "3개 일치 (5,000원) - %d개\n",
@@ -18,21 +15,6 @@ public class View {
     );
     private static final String WINNING_RATE_OUTPUT_STATEMENT = "총 수익율은 %.1f%%입니다.\n";
     private static final String ERROR_OUTPUT_STATEMENT = "[ERROR] %s\n";
-
-    public String getLottoPurchaseAmount() {
-        System.out.println(LOTTO_PURCHASE_AMOUNT_OUTPUT_STATEMENT);
-        return Console.readLine();
-    }
-
-    public String getLottoWinningNumber() {
-        System.out.println(LOTTO_WINNING_NUMBER_OUTPUT_STATEMENT);
-        return Console.readLine();
-    }
-
-    public String getLottoBonusNumber() {
-        System.out.println(LOTTO_BONUS_NUMBER_OUTPUT_STATEMENT);
-        return Console.readLine();
-    }
 
     public void printPurchasedLottoNumber(List<Lotto> tickets) {
         System.out.println(tickets.size() + PURCHASED_LOTTO_NUMBER_OUTPUT_STATEMENT);
